--- app/ns-ui-widgets-category/label/basics/article.md
@@ -1,10 +1,0 @@
-The `Label` widget comes with `text` property.
-
-Setting a label in XML page and how to bind its text property.
-<snippet id='label-xml'/>
-<snippet id='label-text-binding'/>
-<snippet id='label-text-binding-ts'/>
-
-Accessing a label and modifying its properties.
-<snippet id='label-accessing-text'/>
-<snippet id='label-accessing-text-ts'/>
