--- app/ns-ui-widgets-category/label/basics/basics-ts-page.css
@@ -1,5 +1,0 @@
-Label {
-    color: blue;
-    font-size:24;
-    margin: 50;
-}
