--- app/ns-ui-widgets-category/label/basics/basics-ts-page.xml
@@ -1,9 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
-    <Page.actionBar>
-		<ActionBar title="Getting Started"/>
-	</Page.actionBar>
-
-    <GridLayout rows="*">
-        <Label id="lbl" text="{{ title }}" textWrap="true" />
-    </GridLayout>
-</Page>
