--- app/ns-ui-widgets-category/label/code-behind/article.md
@@ -1,3 +1,0 @@
-Creating a `Label` programmatically with `text`.
-<snippet id='label-code-create' />
-<snippet id='label-code-create-ts' />
