+++ app/ns-ui-widgets-category/label/end.md
@@ -1,7 +1,37 @@
+## Tips And Tricks
 
-**API Reference for** [Label Class](http://docs.nativescript.org/api-reference/modules/_ui_label_.html)
+### Creating Label via Code-Behind
+Creating a `Label` programmatically with `text`.
+<snippet id='label-code-create' />
+<snippet id='label-code-create-ts' />
 
-**Native Component**
+## Properties
+
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `letterSpacing`   | `number` | Gets or sets letterSpace style property. |
+| `lineHeight`   | `number` | Gets or sets lineHeight style property. |
+| `text`   | `string` | Gets or sets the Label text. |
+| `textAlignment`   | **_"initial"_**, **_"left"_**, **_"center"_**, **_"right"_** | Gets or sets text-alignment style property. |
+| `textDecoration`   | **_"none"_**, **_"underline"_**, **_"line-through"_**, **_"underline"_**, **_"line-through"_** | Gets or sets text swcoration style property. |
+| `textTransform`   | **_"initial"_**, **_"none"_**, **_"capitalize"_**, **_"uppercase"_**, **_"lowercase"_** | Gets or sets text transform style property. |
+| `textWrap`   | `boolean` | Gets or sets whether the Label wraps text or not. |
+| `whiteSpace`   | **_"initial"_**, **_"normal"_**, **_"nowrap"_** | Gets or sets the white space style. |
+
+## Events
+
+| Name     | Description    |
+|----------|----------------|
+| `textChange`    | Emitted when the label text is changed.|
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/label](http://docs.nativescript.org/api-reference/modules/_ui_label_.html) | `Module` | 
+| [Label](https://docs.nativescript.org/api-reference/classes/_ui_label_.label) | `Class` | 
+
+## Native Component
 
 | Android               | iOS      |
 |:----------------------|:---------|
