+++ app/ns-ui-widgets-category/label/label-page.js
@@ -1,12 +1,11 @@
 const ListViewLinksModel = require("../../links-view-model");
 const link = require("../../link");
 const navigationLinks = [
-    new link("Basics", "ns-ui-widgets-category/label/basics/basics-page"),
-    new link("Code-Behind", "ns-ui-widgets-category/label/code-behind/code-behind-page"),
-    new link("Styling", "ns-ui-widgets-category/label/styling/styling-page")
+    new link("Usage", "ns-ui-widgets-category/label/usage/usage-page"),
+    new link("Code-Behind", "ns-ui-widgets-category/label/code-behind/code-behind-page")
 ];
 const navigationLinksTsc = [
-    new link("Basics", "ns-ui-widgets-category/label/basics/basics-ts-page"),
+    new link("Usage", "ns-ui-widgets-category/label/usage/usage-ts-page"),
     new link("Code-Behind", "ns-ui-widgets-category/label/code-behind/code-behind-ts-page")
 ];
 function onNavigatingTo(args) {
