+++ app/ns-ui-widgets-category/label/overview.md
@@ -1,3 +1,1 @@
-The Label widget provides a text label that shows read-only text.
-<snippet id='require-label'/>
-<snippet id='require-label-ts'/>
+The `Label` widget provides a text label that shows read-only text.
