--- app/ns-ui-widgets-category/label/styling/article.md
@@ -1,5 +1,0 @@
-To style and customize the `Label` widget use the conventional CSS properties and Icon Fonts.
-The example below shows a label with icon (using icon font), several CSS properties.
-
-<snippet id='label-css-styles'/>
-<snippet id='label-styling-xml'/>
