--- app/ns-ui-widgets-category/label/styling/styling-page.css
@@ -1,16 +1,0 @@
-/* >> label-css-styles */
-.my-label {
-    background-color: #bc7474;
-    color: whitesmoke;
-    font-size: 24;
-    padding: 8;
-    text-align: center;
-    vertical-align: middle;
-    width: 90%;
-}
-.border {
-    border-color: darkred;
-    border-radius: 10;
-    border-width: 2;
-}
-/* << label-css-styles */
