--- app/ns-ui-widgets-category/label/styling/styling-page.xml
@@ -1,16 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd">
-    <Page.actionBar>
-		<ActionBar title="Styling"/>
-	</Page.actionBar>
-
-    <GridLayout rows="*">
-    <android>
-        <!-- >> label-styling-xml -->
-        <Label text="&#xff17b; Sample label" class="far my-label border" />
-        <!-- << label-styling-xml -->
-     </android>
-     <ios>
-        <Label text="&#xf179; Sample label!" class="far my-label border" />
-     </ios> 
-    </GridLayout>
-</Page>
