+++ app/ns-ui-widgets-category/label/usage/article.md
@@ -1,0 +1,3 @@
+<snippet id='label-xml'/>
+<snippet id='label-text-binding'/>
+<snippet id='label-text-binding-ts'/>
