+++ app/ns-ui-widgets-category/label/usage/usage-page.xml
@@ -1,11 +1,12 @@
 <Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
     <Page.actionBar>
-		<ActionBar title="Getting Started"/>
+		<ActionBar title="Usage"/>
 	</Page.actionBar>
 
-    <GridLayout rows="*">
+    <GridLayout rows="* *">
         <!-- >> label-xml -->
-        <Label id="lbl" text="{{ title }}" textWrap="true" />
+        <Label text="Lores Ipsum..." textWrap="true"/>
+        <Label row="1" text="{{ title }}" textWrap="true" />
         <!-- << label-xml -->
     </GridLayout>
 </Page>
