+++ app/ns-ui-widgets-category/label/usage/usage-ts-page.ts
@@ -1,0 +1,13 @@
+// >> label-text-binding-ts
+import { EventData, Observable } from "tns-core-modules/data/observable";
+import { Page } from "tns-core-modules/ui/page";
+
+export function onNavigatingTo(args: EventData) {
+    
+    const page = <Page>args.object;
+    const vm = new Observable();
+    vm.set("title", "NativeScript is Awesome");
+    page.bindingContext = vm;
+
+}
+// << label-text-binding-ts
