+++ app/ns-ui-widgets-category/label/usage/usage-ts-page.xml
@@ -1,0 +1,10 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
+    <Page.actionBar>
+		<ActionBar title="Usage"/>
+	</Page.actionBar>
+
+    <GridLayout rows="* *">
+        <Label text="Lores Ipsum..." textWrap="true"/>
+        <Label row="1" text="{{ title }}" textWrap="true" />
+    </GridLayout>
+</Page>
