--- app/ns-ui-widgets-category/list-picker/code-behind/article.md
@@ -1,3 +1,0 @@
-Creating a `ListPicker` programmatically while setting items source and default selected index.
-<snippet id='create-list-picker-code' />
-<snippet id='create-list-picker-code-ts' />
