+++ app/ns-ui-widgets-category/list-picker/end.md
@@ -1,7 +1,26 @@
-**API Reference for** [ListPicker Class](http://docs.nativescript.org/api-reference/modules/_ui_list_picker_.html)
 
-**Native Component**
+## Properties
 
-| Android                | iOS      |
-|:-----------------------|:---------|
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `items`   | `Array<any>` | Gets or set the items collection of the ListPicker. The items property can be set to an array or an object defining length and getItem(index) method. |
+| `selectedIndex`   | `number` | Gets or set the items collection of the ListPicker. The items property can be set to an array or an object defining length and getItem(index) method. |
+
+## Events
+
+| Name     | Description    |
+|----------|----------------|
+| `selectedIndexChange`    | Emitted when the `selectedIndex` is changed.|
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/list-picker](http://docs.nativescript.org/api-reference/modules/_ui_list_picker_.html) | `Module` | 
+| [ListPicker](https://docs.nativescript.org/api-reference/classes/_ui_list_picker_.listpicker) | `Class` | 
+
+## Native Component
+
+| Android               | iOS      |
+|:----------------------|:---------|
 | [android.widget.NumberPicker](http://developer.android.com/reference/android/widget/NumberPicker.html) | [UIPickerView](https://developer.apple.com/library/prerelease/ios/documentation/UIKit/Reference/UIPickerView_Class/index.html) |
