+++ app/ns-ui-widgets-category/list-picker/overview.md
@@ -1,5 +1,1 @@
-Spinner type module with listed options.
-Using a [ListPicker]() requires the `tns-core-modules/ui/list-picker` module.
-
-<snippet id='require-list-picker'/>
-<snippet id='require-list-picker-ts'/>
+The `ListPicker` is a spinner type component for listing options.
