+++ app/ns-ui-widgets-category/list-picker/tips-and-tricks/tips-and-tricks-page.xml
@@ -1,0 +1,17 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
+    <Page.actionBar>
+		<ActionBar title="Getting Started"/>
+	</Page.actionBar>
+    <StackLayout>
+        <!-- >> list-picker-key-value-pairs-xml -->
+        <ListPicker 
+            loaded="onListPickerLoaded"
+            items="{{ items }}"
+            textField="name"
+            valueField="role"
+            selectedIndex="{{ index }}"
+            selectedValue="{{ selectedItem }}">
+        </ListPicker>
+        <!-- << list-picker-key-value-pairs-xml -->
+    </StackLayout>
+</Page>
