+++ app/ns-ui-widgets-category/list-picker/usage/usage-page.js
@@ -1,10 +1,8 @@
 // >> list-picker-code-behind
 const fromObject = require("tns-core-modules/data/observable").fromObject;
-
+const years = [1980, 1990, 2000, 2010, 2020];
 function onNavigatingTo(args) {
-    
-    const years = [1980, 1990, 2000, 2010, 2020];
-
+console.log(years)
     const page = args.object;
     const vm = fromObject({
         years: years
@@ -18,7 +16,7 @@ function onListPickerLoaded(fargs) {
     const listPickerComponent = fargs.object;
     listPickerComponent.on("selectedIndexChange", (args) => {
         const picker = args.object;
-        console.log(`index: ${picker.selectedIndex}; item" ${this.years[picker.selectedIndex]}`);
+        console.log(`index: ${picker.selectedIndex}; item" ${years[picker.selectedIndex]}`);
     });
 }
 exports.onListPickerLoaded = onListPickerLoaded;
