+++ app/ns-ui-widgets-category/list-picker/usage/usage-ts-page.ts
@@ -2,10 +2,10 @@
 import { EventData, fromObject } from "tns-core-modules/data/observable";
 import { Page } from "tns-core-modules/ui/page";
 import { ListPicker } from "tns-core-modules/ui/list-picker/list-picker";
-
+const years = [1980, 1990, 2000, 2010, 2020];
 export function onNavigatingTo(args: EventData) {
     
-    const years = [1980, 1990, 2000, 2010, 2020];
+    
 
     const page = <Page>args.object;
     const vm = fromObject({
@@ -19,7 +19,7 @@ export function onListPickerLoaded(fargs) {
     const listPickerComponent = fargs.object;
     listPickerComponent.on("selectedIndexChange", (args: EventData) => {
         const picker = <ListPicker>args.object;
-        console.log(`index: ${picker.selectedIndex}; item" ${this.years[picker.selectedIndex]}`);
+        console.log(`index: ${picker.selectedIndex}; item" ${years[picker.selectedIndex]}`);
     });
 }
 // << list-picker-code-behind-ts
